refactor: split tasks
