+++ backend/frege/celery_app.py
@@ -21,4 +21,10 @@ app.config_from_object("django.conf:settings", namespace="CELERY")
 # Load task modules from all registered Django app configs.
 app.autodiscover_tasks()
 
+app.autodiscover_tasks(
+    [
+        "frege.repositories.tasks",
+    ],
+)
+
 app.conf.broker_connection_retry_on_startup = True
